Jest: Refactor list formatter
